Added dirty form submit to checklists (see description)
To allow a form to submit only changed fields, add the class 'dirty-submit' to the form.

If the form dynamically sets the value of hidden input fields using val(), ensure that change() is also triggered.

If the form dynamically adds fields, ensure that the class 'changed' is added to all input fields (including hidden) that were added.