Adds ErrorMailer with mobile_debug_report
This will send an email to engineering@prosafet.com when a user sends a debug report from the ProSafeT app. This email includes a description of the issue, the user's name and id, the airline, device information, and app version. It also includes a json file attachments with a dump of all app data, including cached data and app state.